move joker fn
