+++ src/algorithm-v3.ts
@@ -1,37 +1,11 @@
 import type { Tile } from "./tiles/tile";
-import { NumberedTile, FreeJokerTile, BoundJokerTile } from "./tiles/tile";
+import { BoundJokerTile } from "./tiles/tile";
 import type { TileOnBoard } from "./tiles/tile";
 import type { GameState } from "./game-state";
 import { MELD_THRESHOLD } from "./constants";
 import { getTileGroups } from "./gameplay/get-tile-groups";
 import { getTileRuns } from "./gameplay/get-tile-runs";
 import { getScore } from "./tiles/scoring";
-
-export const substituteJokers = (gameState: GameState): GameState => {
-  const { board, hand } = gameState;
-  const numberedTilesInHand: NumberedTile[] = hand.filter(
-    (t) => t instanceof NumberedTile,
-  );
-  const addedFreeJokerTiles: FreeJokerTile[] = [];
-  for (const set of board) {
-    const boundJokerTiles = set.filter((t) => t instanceof BoundJokerTile);
-    for (const boundJokerTile of boundJokerTiles) {
-      for (const tile of numberedTilesInHand) {
-        if (boundJokerTile.matches(tile)) {
-          const indexOfNumberedTileInHand = numberedTilesInHand.indexOf(tile);
-          numberedTilesInHand.splice(indexOfNumberedTileInHand, 1);
-          addedFreeJokerTiles.push(new FreeJokerTile());
-          const indexOfBoundJokerTileInSet = set.indexOf(boundJokerTile);
-          set[indexOfBoundJokerTileInSet] = tile;
-        }
-      }
-    }
-  }
-  return {
-    board,
-    hand: [...numberedTilesInHand, ...addedFreeJokerTiles],
-  };
-};
 
 export const findBestMeldingPlay = (gameState: GameState): GameState => {
   const { hand, board } = gameState;
